Support playback of `.strm` files
Add new `StreamType`: `.strm`

Set correct `StreamType` and `URL` for `.strm`

Send the `sendStartReport` immediately when playback starts, because direct play of `.strm` files reset the playback to `00:00:00` at start of playback. Context: in a `directPlay` state, the server will start playback progress tracking from `00:00:00`, until the progressReport is sent. This can cause media to have an incorrect playback progress status, if the `sendStartReport` task is never executed (e.g. the task in DispatchQueue doesn't execute for whatever reason)

Set `AudioActionButton` and `SubtitleActionButton` correctly when playback starts. Context: When playback starts, the value of the `videoPlayerManager` is not set to the correct index of audio and subtitle tracks. A better solution would be to update `videoPlayerManager` with the currently selected index values from `VideoPlayerViewModel`.

Allow `.strm` playback on NativePlayer