[tvOS][Bug] Fixes season selector only showing a single season.
Upon initial view load, only the first season is loaded into `menuSections` when the episode list is resolved.

This change stores the list of all seasons, which contain their display name.  When changing the series selector, `menuSections` is updated with season details.

Also when changing seasons in the picker, the containing scroll view scrolled to the top of the page.  This was jarring.  This change also removes that behavior.