Several Changes:
Migrate over to `Material` for`PillHStack`.
Migrate Close & Menu buttons in Navigation with `TintedMaterial`
Migrate filters over to `TintedMaterial` & update `TintedMaterial` to have an optional `isPressed`
Remove `isPressed` since the only usages are in `TintedMaterial`
Remove the static rounding on `TintedMaterial` to instead go in a case by case bases so I can do circles for some and rounded rectangles for others.